Remove redundant venv installation step
Dropped an unnecessary block in the install script that checked and installed `venv`. The installation of `venv` is now implicitly handled elsewhere, simplifying the installation process and reducing the script's length. Additionally, the manual installation of `virtualenv` in the Darwin OS section has been removed as it's included in the common `pip3 install` command.